refactor(dtm): remove obsolete stats module
Closes #296
